Module pallet_xc_asset_config::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

change_existing_asset_location

Change the xcm type mapping for a given asset Id. The new asset type will inherit old units per second value.

register_asset_location

Register new asset location to asset Id mapping.

remove_asset

Removes all information related to asset, removing it from XCM support.

remove_payment_asset

Removes asset from the set of supported payment assets.

set_asset_units_per_second

Change the amount of units we are charging per execution second for a given AssetLocation.